expose Input[] and Output[] type generics
