+++ docs/source/dsl.rst
@@ -2,6 +2,3 @@ kfp.dsl
 ==========================
 
 .. automodule:: kfp.dsl
-
-    .. autodata:: Input
-    .. autodata:: Output
