+++ sdk/python/kfp/components/types/type_annotations.py
@@ -22,13 +22,6 @@ from typing import Type, TypeVar, Union
 from kfp.components.types import type_annotations
 from kfp.components.types import type_utils
 
-try:
-    from typing import Annotated
-except ImportError:
-    from typing_extensions import Annotated
-
-T = TypeVar('T')
-
 
 class OutputPath:
     """Type annotation used in component definitions for indicating a parameter
@@ -127,59 +120,6 @@ class OutputAnnotation():
     """Marker type for output artifacts."""
 
 
-Input = Annotated[T, InputAnnotation]
-Input.__doc__ = """Type generic used to represent an input artifact of type ``T``, where ``T`` is an artifact class.
-
-Use ``Input[Artifact]`` or ``Output[Artifact]`` to indicate whether the enclosed artifact is a component input or output.
-
-Args:
-    T: The type of the input artifact.
-
-Example:
-  ::
-
-    @dsl.component
-    def artifact_producer(model: Output[Artifact]):
-        with open(model.path, 'w') as f:
-            f.write('my model')
-
-    @dsl.component
-    def artifact_consumer(model: Input[Artifact]):
-        print(model)
-
-    @dsl.pipeline(name='my-pipeline')
-    def my_pipeline():
-        producer_task = artifact_producer()
-        artifact_consumer(model=producer_task.output)
-"""
-
-Output = Annotated[T, OutputAnnotation]
-Output.__doc__ = """A type generic used to represent an output artifact of type ``T``, where ``T`` is an artifact class. The argument typed with this annotation is provided at runtime by the executing backend and does not need to be passed as an input by the pipeline author (see example).
-
-Use ``Input[Artifact]`` or ``Output[Artifact]`` to indicate whether the enclosed artifact is a component input or output.
-
-Args:
-    T: The type of the output artifact.
-
-Example:
-  ::
-
-    @dsl.component
-    def artifact_producer(model: Output[Artifact]):
-        with open(model.path, 'w') as f:
-            f.write('my model')
-
-    @dsl.component
-    def artifact_consumer(model: Input[Artifact]):
-        print(model)
-
-    @dsl.pipeline(name='my-pipeline')
-    def my_pipeline():
-        producer_task = artifact_producer()
-        artifact_consumer(model=producer_task.output)
-"""
-
-
 def is_artifact_annotation(typ) -> bool:
     if not hasattr(typ, '__metadata__'):
         return False
@@ -220,6 +160,9 @@ def get_io_artifact_annotation(typ):
         return None
 
     return typ.__metadata__[0]
+
+
+T = TypeVar('T')
 
 
 def maybe_strip_optional_from_annotation(annotation: T) -> T:
